In the model, Messages form a FIFO queue.
Prior to this change, an Actor that received multiple messages and had
not yet handled them could choose any available message to handle,
leading to less determinism than we want to model for Actor messaging.

After this change, Messages are modeled as a FIFO queue on the Actor,
and the Actor can only process the next immediate message in the queue.